Use FormAction instead of UnknownAction hook, Security fix
This includes an important security fix for CSRF.

Rewrite code to base on FormAction class instead
of the deprecated UnknownAction hook.

Provides one action 'delete_page_permanently' only.

Credits - user saper helped to do the code.

Some basic postgress support. Since postgress dosen't have the searchindex
table lets disable it for this db type.

Fix images.

Bug: T117280
Bug: T122540
Bug: T119304

+++ ActionDeletePagePermanently.php
@@ -1,94 +1,62 @@
 <?php
 
-class DeletePagesForGood {
+class ActionDeletePagePermanently extends FormAction {
 
-	function __construct() {
-		global $wgHooks;
-
-		$wgHooks['SkinTemplateNavigation::Universal'][] = array(
-			&$this,
-			'AddSkinHook'
-		);
-
-		$wgHooks['UnknownAction'][] = array(
-			&$this,
-			'AddActionHook'
-		);
-	}
-
-	function AddSkinHook( SkinTemplate &$sktemplate, array &$links ) {
-		global $wgRequest, $wgTitle, $wgUser, $wgDeletePagesForGoodNamespaces;
-
-		if ( !$wgUser->isAllowed( 'deleteperm' ) ) {
+	public static function AddSkinHook( SkinTemplate &$sktemplate, array &$links ) {
+		if ( !$sktemplate->getUser()->isAllowed( 'deleteperm' ) ) {
 
 			return false;
 		}
 
-		$action = $wgRequest->getText( 'action' );
+		$title = $sktemplate->getRelevantTitle();
+		$action = self::getActionName( $sktemplate );
 
-		# Special pages can not be deleted (special pages have no article id anyway).
-		if ( $wgTitle->getArticleID() != 0
-			&& isset( $wgDeletePagesForGoodNamespaces[$wgTitle->getNamespace()] )
-			&& $wgDeletePagesForGoodNamespaces[$wgTitle->getNamespace()] == true
-			&& $wgTitle->getNamespace() != NS_SPECIAL
+		if ( self::canDeleteTitle( $title ) ) {
+			$links['actions']['delete_page_permanently'] = array(
+				'class' => ( $action === 'delete_page_permanently' ) ? 'selected' : false,
+				'text' => $sktemplate->msg( 'deletepagesforgood-delete_permanently' )->text(),
+				'href' => $title->getLocalUrl( 'action=delete_page_permanently' )
+			);
+		}
+
+		return true;
+	}
+
+	public function getName() {
+		return 'delete_page_permanently';
+	}
+
+	public function getDescription() {
+		return '';
+	}
+
+	public static function canDeleteTitle( Title $title ) {
+		global $wgDeletePagesForGoodNamespaces;
+
+		if ( $title->exists() && $title->getArticleID() !== 0 &&
+			$title->getDBkey() !== '' &&
+			$title->getNamespace() !== NS_SPECIAL &&
+			isset( $wgDeletePagesForGoodNamespaces[ $title->getNamespace() ] ) &&
+			$wgDeletePagesForGoodNamespaces[ $title->getNamespace() ]
 		) {
-			$links['actions']['ask_delete_page_permanently'] = array(
-				'class' => ( $action == 'ask_delete_page_permanently' ) ? 'selected' : false,
-				'text' => wfMessage( 'deletepagesforgood-delete_permanently' )->text(),
-				'href' => $wgTitle->getLocalUrl( 'action=ask_delete_page_permanently' )
-			);
-		}
-
-		return true;
-	}
-
-	function AddActionHook( $action, $wgArticle ) {
-		global $wgOut, $wgUser, $wgDeletePagesForGoodNamespaces;
-
-		if ( !$wgUser->isAllowed( 'deleteperm' ) ) {
-			throw new PermissionsError( 'deleteperm' );
-		}
-
-		# Print a form to approve deletion
-		if ( $action == 'ask_delete_page_permanently' ) {
-
-			$action = $wgArticle->getTitle()->getLocalUrl( 'action=delete_page_permanently' );
-			$wgOut->addHTML( "<form id='ask_delete_page_permanently' method='post' action=\"$action\">
-				<table>
-						<tr>
-							<td>" . wfMessage( 'deletepagesforgood-ask_deletion' )->text() . "</td>
-						</tr>
-						<tr>
-							<td><input type='submit' name='submit' value=\"" .
-								wfMessage( 'deletepagesforgood-yes' )->text() . "\" />
-							</td>
-						</tr>
-				</table></form>"
-			);
-			return false;
-		} elseif ( $action == 'delete_page_permanently' ) {
-			# Perform actual deletion
-			$ns = $wgArticle->mTitle->getNamespace();
-			$t = $wgArticle->mTitle->getDBkey();
-			$id = $wgArticle->mTitle->getArticleID();
-
-			if ( $t == '' || $id == 0 || $wgDeletePagesForGoodNamespaces[$ns] != true
-				|| $ns == NS_SPECIAL
-			) {
-				$wgOut->addHTML( wfMessage( 'deletepagesforgood-del_impossible' )->escaped() );
-				return false;
-			}
-
-			$this->deletePermanently( $wgArticle->mTitle );
-			$wgOut->addHTML( wfMessage( 'deletepagesforgood-del_done' )->escaped() );
+			return true;
+		} else {
 			return false;
 		}
-
-		return true;
 	}
 
-	function deletePermanently( $title ) {
-		global $wgOut;
+	public function onSubmit( $data ) {
+
+		if ( self::canDeleteTitle( $this->getTitle() ) ) {
+			$this->deletePermanently( $this->getTitle() );
+			return true;
+		} else {
+			# $output->addHTML( $this->msg( 'deletepagesforgood-del_impossible' )->escaped() );
+			return array( 'deletepagesforgood-del_impossible' );
+		}
+	}
+
+	public function deletePermanently( Title $title ) {
 
 		$ns = $title->getNamespace();
 		$t = $title->getDBkey();
@@ -112,8 +80,10 @@ class DeletePagesForGood {
 		# delete language link...
 		$dbw->delete( 'langlinks', array( 'll_from' => $id ), __METHOD__ );
 
-		# delete search index...
-		$dbw->delete( 'searchindex', array( 'si_page' => $id ), __METHOD__ );
+		if ( !$GLOBALS['wgDBtype'] == "postgres" ) {
+			# delete search index...
+			$dbw->delete( 'searchindex', array( 'si_page' => $id ), __METHOD__ );
+		}
 
 		# Delete restrictions for the page
 		$dbw->delete( 'page_restrictions', array( 'pr_page' => $id ), __METHOD__ );
@@ -197,7 +167,7 @@ class DeletePagesForGood {
 		/*
 		 * If an image is beeing deleted, some extra work needs to be done
 		 */
-		if ( $ns == NS_IMAGE ) {
+		if ( $ns == NS_FILE ) {
 
 			$file = wfFindFile( $t );
 
@@ -214,17 +184,17 @@ class DeletePagesForGood {
 						unlink( $path );
 					}
 					catch ( Exception $e ) {
-						$wgOut->addHTML( $e->getMessage() );
+						return $e->getMessage();
 					}
 				}
 
-				$path = $file->getPath();
+				$path = $file->getLocalRefPath();
 
 				try {
 					$file->purgeThumbnails();
 					unlink( $path );
 				} catch ( Exception $e ) {
-					$wgOut->addHTML( $e->getMessage() );
+					return $e->getMessage();
 				}
 			}
 
@@ -245,5 +215,40 @@ class DeletePagesForGood {
 			$linkCache = LinkCache::singleton();
 			$linkCache->clear();
 		}
+		return true;
+	}
+
+	/**
+	 * Returns the name that goes in the \<h1\> page title
+	 *
+	 * @return string
+	 */
+	protected function getPageTitle() {
+		return $this->msg( 'deletepagesforgood-deletepagetitle', $this->getTitle()->getPrefixedText() );
+	}
+
+	protected function alterForm( HTMLForm $form ) {
+
+		$title = $this->getTitle();
+		$output = $this->getOutput();
+
+		$output->addBacklinkSubtitle( $title );
+		$form->addPreText( $this->msg( 'confirmdeletetext' )->parseAsBlock() );
+
+		$form->addPreText(
+			$this->msg( 'deletepagesforgood-ask_deletion' )->parseAsBlock()
+		);
+
+		$form->setSubmitTextMsg( 'deletepagesforgood-yes' );
+	}
+
+	public function getRestriction() {
+		return 'deleteperm';
+	}
+
+	public function onSuccess() {
+		$output = $this->getOutput();
+		$output->addHTML( $this->msg( 'deletepagesforgood-del_done' )->escaped() );
+		return false;
 	}
 }
